backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database
sqlalchemy[asyncio]>=2.0.46
asyncpg>=0.30.0
alembic>=1.14.0

# Auth
PyJWT>=2.10.0
pwdlib[argon2,bcrypt]>=0.2.0

# File Handling
python-multipart>=0.0.20
aiofiles>=24.1.0
Pillow>=12.1.0

# HTTP Client
httpx>=0.27.0

# Billing (optional, aktivierbar per Admin-Setting)
stripe>=12.0.0

# HTML Parsing (Website Importer)
beautifulsoup4>=4.12.0

# YAML
PyYAML>=6.0.0

# Content Quality Analysis
textstat>=0.7.4

# Utilities
nanoid>=2.0.0
email-validator>=2.0.0
